Rebuild TestPool timeout handling
Rebuild timeout handling not using the Timeout module but a
ConditionalVariable for signaling when new messages arrived.

Therefore the consumer execute is not longer directly involved in the
timeout except when waiting on multiple messages. It still would only
abort when waiting on message to arive, if newly enqueued expected
messages arive the timeout won't even be touched.