Allow passing AMQP connection parameters as hash
This commit changes the configuration parsing code to correctly accept
AMQP connection parameters as a hash and add support for slash-less
vhost arguments. Previously an error was raised when trying to assign a
slash-less vhost as a URI path. The actual `uri` configuration is still
parsed in the previous non-AMQP variant using the actual URI path.

This commit allows a configuration as the following:

    production:
      uri: amqp://user@host/vhost
      pass: secret
      vhost: real_vhost
      port: 1234

With the above configuration MSGR will try to connect to a broker at
`host:1234`, authenticate with `user` and `secret` and using the
vhost `real_vhost`.

- Fix #13