Drop autostart
Starting the client in-process has proven to be flaky, hard to target
(e.g. rake tasks) and not scalable. This CL completly removed the
autostart option.

Client that really want this behavior are required to start the client
themself, e.g. in an initializer with appropiate checks.

This also removed a lot of complex options parsing code. Config keys are
taken as they are now and used as bool values. Changing the config path
with `app.config.msgr.rabbitmq_config` has been dropped without
replacement.