Please Rubocop: Remove unused method
Rubocop is complaining about the Boolean argument. Changing this would
be a BC break, technically. Since the method does not seem to be used
and isn't really public API, I decided to remove it. (Also a BC break,
technically.) The two similar methods `#ack` and `#nack` are being
delegated to from the `Message` class - `#reject` is not.